+++ ros_gz_bridge/ros_gz_bridge/mappings.py
@@ -74,6 +74,7 @@ MAPPINGS = {
         Mapping('JointWrench', 'JointWrench'),
         Mapping('Light', 'Light'),
         Mapping('LogicalCameraImage', 'LogicalCameraImage'),
+        Mapping('LogPlaybackStatistics', 'LogPlaybackStatistics'),
         Mapping('MaterialColor', 'MaterialColor'),
         Mapping('ParamVec', 'Param'),
         Mapping('ParamVec', 'Param_V'),
@@ -81,6 +82,7 @@ MAPPINGS = {
         Mapping('StringVec', 'StringMsg_V'),
         Mapping('TrackVisual', 'TrackVisual'),
         Mapping('VideoRecord', 'VideoRecord'),
+        Mapping('WorldStatistics', 'WorldStatistics'),
     ],
     'rosgraph_msgs': [
         Mapping('Clock', 'Clock'),