+++ ros_gz_sim/launch/gz_server.launch.py
@@ -37,6 +37,12 @@ def generate_launch_description():
     declare_use_composition_cmd = DeclareLaunchArgument(
         'use_composition', default_value='False',
         description='Use composed bringup if True')
+    declare_initial_sim_time_cmd = DeclareLaunchArgument(
+        'initial_sim_time', default_value='0.0',
+        description='The initial simulation time')
+    declare_verbosity_level_cmd = DeclareLaunchArgument(
+        'verbosity_level', default_value='4',
+        description='The verbosity level of the Gazebo server (0=FATAL, 4=DEBUG)')
 
     gz_server_action = GzServer(
         world_sdf_file=LaunchConfiguration('world_sdf_file'),
@@ -44,6 +50,8 @@ def generate_launch_description():
         container_name=LaunchConfiguration('container_name'),
         create_own_container=LaunchConfiguration('create_own_container'),
         use_composition=LaunchConfiguration('use_composition'),
+        initial_sim_time=LaunchConfiguration('initial_sim_time'),
+        verbosity_level=LaunchConfiguration('verbosity_level'),
     )
 
     # Create the launch description and populate
@@ -55,6 +63,8 @@ def generate_launch_description():
     ld.add_action(declare_container_name_cmd)
     ld.add_action(declare_create_own_container_cmd)
     ld.add_action(declare_use_composition_cmd)
+    ld.add_action(declare_initial_sim_time_cmd)
+    ld.add_action(declare_verbosity_level_cmd)
     # Add the gz_server action
     ld.add_action(gz_server_action)